pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "erpclasp"
version = "0.1.0"
description = "CLI to pull, push, diff, and watch ERPNext Server Scripts via Frappe REST API"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "erpclasp contributors" }]
keywords = ["erpnext", "frappe", "cli", "server-script"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.12.0",
    "requests>=2.31.0",
    "watchdog>=4.0.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.4.0", "mypy>=1.9.0", "build>=1.0.0", "twine>=5.0.0"]

[project.urls]
Homepage = "https://github.com/your-org/erpclasp"
Repository = "https://github.com/your-org/erpclasp"

[project.scripts]
erpclasp = "erpclasp.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["erpclasp*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
